fix: Reuse already defined schema and add Paginated Schema
